Move Erased* to kas_core::message
